moved: Club type 선언경로 승격
